# Optimize takeWhile and takeWhile1