@bgamari 2016-04-06
There are a few open questions here,

 * Do we want to offer this function at all? Perhaps the cost of always inlining `takeWhile` isn't as much as I suspect (although testing this is rather tricky) and we should just mark it as `INLINE`
 * Perhaps a more descriptive name like `takeWhileInlined` would be appropriate
 * If the cost of using the inlined version inappropriately is high then perhaps it should be set apart from the primary interface (either in a different Haddock section or module) to avoid users assuming "ooh, inlining means always faster, right?"