@Yuras 2016-04-06
I think `takeWhile` should be `INLINE` on itself. Most likely it was just incidentally overlooked. Note that `Data.Attoparsec.ByteString.takeWhile` is already inlined.

Also `takeWhile` is suboptimal on itself, I'd expect ~20% speedup by separating a slow path (not enough input) from a fast path, like it is done in scanner: http://haddock.stackage.org/nightly-2016-04-06/scanner-0.2/src/Scanner-Internal.html#takeWhile