@bgamari 2016-04-06
@Yuras, fair enough. I honestly don't have a good sense for how much this code bloat would hurt performance. The code looks quite large, but I haven't yet seen a case where inlining hurts.

Also, note that `Data.Attoparsec.ByteString.takeWhile1` is not inlined. Perhaps this is another oversight?