@bgamari 2016-04-06
Here's a new set of patches performing the refactoring suggested by @yuras and inlining always `takeWhile` and `takeWhile1`.