@bgamari 2016-04-08
Closed in favor of #119 (every time I open a pull request from my `master` branch I eventually regret it).